+++ src/main/java/io/scalecube/socketio/ServerBootstrapFactory.java
@@ -1,0 +1,12 @@
+package io.scalecube.socketio;
+
+import io.netty.bootstrap.ServerBootstrap;
+
+/**
+ * @author Anton Kharenko
+ */
+public interface ServerBootstrapFactory {
+
+  ServerBootstrap createServerBootstrap();
+
+}