+++ src/main/java/io/scalecube/socketio/pipeline/XHRPollingHandler.java
@@ -36,11 +36,11 @@ public class XHRPollingHandler extends ChannelInboundHandlerAdapter {
   private final Logger log = LoggerFactory.getLogger(getClass());
 
   private final String connectPath;
-  private final String headerClientIpAddressName;
+  private final String remoteAddressHeader;
 
-  public XHRPollingHandler(final String handshakePath, final String headerClientIpAddressName) {
+  public XHRPollingHandler(final String handshakePath, final String remoteAddressHeader) {
     this.connectPath = handshakePath + TransportType.XHR_POLLING.getName();
-    this.headerClientIpAddressName = headerClientIpAddressName;
+    this.remoteAddressHeader = remoteAddressHeader;
   }
 
   @Override
@@ -59,7 +59,7 @@ public class XHRPollingHandler extends ChannelInboundHandlerAdapter {
         final String origin = PipelineUtils.getOrigin(req);
 
         if (HttpMethod.GET.equals(requestMethod)) {
-          SocketAddress clientIp = PipelineUtils.getHeaderClientIPParamValue(req, headerClientIpAddressName);
+          SocketAddress clientIp = PipelineUtils.getHeaderClientIPParamValue(req, remoteAddressHeader);
 
           // Process polling request from client
           final ConnectPacket packet = new ConnectPacket(sessionId, origin);