+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes
 * Make configurable WebSocket max frame size
 * Remove memoizer from SessionStorage
 * Upgrade netty to 4.0.33.Final
+* Made ServerBootstrap injectable
 
 2.0.1 / 2015-09-10
 -----------------------